fourth

+++ http.py
@@ -4,53 +4,33 @@ from datetime import *
 import os.path, time
 from urllib.parse import *
 import shutil
-import file_format
+import config
+import getpass
+import os
 
-def curr_time():
+def curr_time():#returns the current time in the required format
 		t = datetime.now()
 		t = t.strftime("%a, %d %b %Y %H:%M:%S")
 		return t
 
-def modified(file_path):
+def modified(file_path):#gets the modified time of file 
 	t = os.path.getmtime(file_path)
 	t = time.strftime("%a, %d %b %Y %H:%M:%S", time.localtime(t)) + " GMT"
 	return t
 
-class Response:
-	status = {
-		200 : "Ok",
-		201	: "Created",
-		202	: "Accepted",
-		204 : "No Content",
-		304 : "Not Modified",
-		400 : "Bad Request",
-		404 : "Not Found",
-		411 : "Length required",
-		412 : "Precondition Failed",
-		414 : "URI too long",
-		415 : "Unsupported media Type",
-		505 : "HTTP version not supported",
-	}
-
-	file_format = file_format.FORMAT
-
-	html_string_for_post = '''<!DOCTYPE html>
-<html>
-<head>
-	<title>form test</title>
-</head>
-<body>
-	<h1> Resource created!</h1>
-</body>
-</html>'''
-
+class Response:		#Class defined for headers manipulation of response object ; To be sent to the client
+	status = config.status_codes
+	file_format = config.FORMAT
+	html_string_for_post = config.post_string
 
 	def __init__(self, request):
 		self.response = "HTTP/1.1 " + "\r\n"
 		self.response += "Date: " + curr_time() + "\r\n"
 		self.response += "Server: Local" + "\r\n"
 
-	def handle_200(self, request):
+	'''All the methods starting with 'handle_status' are used to handle the headers required for the 
+	given status code. Only 'handle_4xx' handles the response status codes starting with 4.'''
+	def handle_200(self, request): 
 		filename, file_extension = os.path.splitext(request["resource"])
 		file_format_header = self.file_format.get(file_extension, None)
 		if not file_format_header:
@@ -101,9 +81,9 @@ class Response:
 		except:
 			print("The status file could not be opened")
 		entity = file.read()
+		file.close()
 		l = len(entity)
 		self.response += "Content-Length: " + str(l) + "\r\n"
-		self.response += "Connection: " + request["Connection"] + "\r\n"
 		self.response += "Content-type: text/html" + "\r\n\r\n"
 		self.response += entity
 		return self.response
@@ -111,6 +91,19 @@ class Response:
 	def handle_405(self, request):
 		index = self.response.find("\r\n")
 		self.response = self.response[:index] + str(stat) + " " + self.status[stat] + self.response[index:]
+		return self.response
+
+	def handle_501(self, request):
+		index = self.response.find("\r\n")
+		self.response = self.response[:index] + "501 " + self.status[501] + self.response[index:]
+		file = open("501.html", "r")
+		entity = file.read()
+		file.close()
+		l = len(entity)
+		self.response += "Content-Length: " + str(l) + "\r\n"
+		self.response += "Content-type: text/html" + "\r\n\r\n"
+		self.response += entity
+		return self.response
 
 	def handle_505(self, request):
 		index = self.response.find("\r\n")
@@ -118,92 +111,108 @@ class Response:
 		return self.response
 
 
-class server:
-	methods = ["GET", "POST", "HEAD", "PUT", "DELETE", "TRACE"]
+class server: #Creates server object ready to listen to a given number of clients at the same time.
+
+	methods = config.methods
+	authorization_pass = config.auth_password
 
 	def __init__(self, address):
 		try:
-			server_sock = socket(AF_INET, SOCK_STREAM)
+			server_sock = socket(AF_INET, SOCK_STREAM)#create a TCP socket
 		except OSError:
 			print("The socket couldn't be created")
-		server_sock.bind(address)
+		server_sock.bind(address)  #Binds the socket to the given address
 		self.start_server(server_sock)
 
 	def start_server(self, server_sock):
 		print("The server has started running..\n" + "Press ctrl + c to quit")
-		server_sock.listen(40)
+		server_sock.listen(100) #Starts listening for tcp connection
 		while True:
-			new_connection, recv_addr = server_sock.accept()
+			new_connection, recv_addr = server_sock.accept() #Connect to the requesting client
+			'''The following line creates a thread for each new requesting client'''
 			client = Thread(group = None, target = self.run, kwargs = {'new_connection' : new_connection, 'recv_addr' : recv_addr})
-			client.start()
-
+			client.start() #Start the thread dedicated to requests only from the given client
 
 	def run(self, new_connection, recv_addr):
-		BUFF_SIZE = 8192
+		BUFF_SIZE = config.buffer_size
 		while True:
 			recv_message = new_connection.recv(BUFF_SIZE)
-			k = recv_message.split(b"\r\n\r\n")
+			k = recv_message.split(b"\r\n\r\n", 1) #split the headers from the body
 			recv_decoded_message = k[0].decode()
 			if len(k) > 1:
 				body = k[1]
-			request = self.handle_request(recv_decoded_message)
+			'''Processess the request and returns it in the form of dictionary; Key: header
+			Value: Value assigned to the header.'''
+			request, status = self.handle_request(recv_decoded_message) 
+			if not status: #If no problem with headers.
+				'''If the body is not fully received, the recv function is called repeatedly to accumulate
+				the fragments in a list and then joined together.'''
+				if request['version'] and request['method'] == "PUT" :
+					fragments = []
+					fragments.append(body)
+					while True:
+						data = new_connection.recv(BUFF_SIZE)
+						fragments.append(data)
+						if len(data) < BUFF_SIZE:
+							break
 
-			if not self.is_good(request):
-				message = Response(request)
-				response = message.handle_4xx(request, 400)
-				new_connection.send(response.encode())
-
-			if request['version'] and request['method'] == "PUT" :
-				fragments = [body]
-				while True:
-					data = new_connection.recv(BUFF_SIZE)
-					fragments.append(data)
-					if len(data) < BUFF_SIZE:
-						break
-
-				body = b"".join(fragments)
-				self.put(request, body, new_connection)
-
-			if request['version'] and request['method'] == "GET":
-				self.get(request, new_connection)
-			elif request['version'] and request['method'] == "HEAD":
-				self.head(request, new_connection)
-			elif request['version'] and request['method'] == "POST":
-				self.post(request, body.decode(), new_connection)
-			elif request['version'] and request['method'] == "DELETE":
-				self.delete(request, new_connection)
+					body = b"".join(fragments)
+					self.put(request, body, new_connection)
 				
+				elif request['version'] and request['method'] == "TRACE":
+					self.trace(request, k[0], new_connection)
+				elif request['version'] and request['method'] == "GET":
+					self.get(request, new_connection)
+				elif request['version'] and request['method'] == "HEAD":
+					self.head(request, new_connection)
+				elif request['version'] and request['method'] == "POST":
+					self.post(request, body.decode(), new_connection)
+				elif request['version'] and request['method'] == "DELETE":
+					self.delete(request, new_connection)
+			else: #If any defect in headers
+				message = Response(request)
+				if status >= 400 and status < 500:
+					response = message.handle_4xx(request, status)
+					new_connection.send(response.encode())
+				elif status == 501:
+					response = message.handle_501(request)
+					new_connection.send(response.encode())
+				elif status == 505:
+					response = message.handle_505(request)
+					new_connection.send(response.encode())
+
 		new_connection.close()
 
-
+		'''Following is the function to do string processing to return the request dictionary. ''' 
 	def handle_request(self, recv_message):
 		request = {}
+		status = None
 		recv_message = recv_message.split("\r\n")
 		first_line = recv_message[0].split(" ")
 		request['method'] = first_line[0]
+		if request['method'] not in self.methods:
+			status = 501
 		if len(first_line) > 1:
-			request['resource'] = self.handle_URI(request, first_line[1])
-			if first_line[2] == "HTTP/1.1":
-				request['version'] = True
+			if len(first_line[1]) > config.MAX_URI_LEN:
+				status = 414
+			else:
+				if first_line[2] == "HTTP/1.1":
+					request['version'] = True
+				else:
+					request['version'] = False
+			request['resource'] = self.handle_URI(request, first_line[1]) #Handles the URL
+			
 		else:
 			request['version'] = False
+			status = 400	
 		for line in recv_message[1:]:
 			i = line.split(": ")
 			if len(i) == 2:
 				request[i[0]] = i[1]
 
-		return request
+		return request, status
 
-	def is_good(self, request):
-		if not request['version']:
-			return False
-		if request['method'] not in self.methods:
-			return False
-		host = request.get("Host", None)
-		if not host:
-			return False
-		return True
-
+	'''Following is the function to make sense of the URL. It prints the get parameters on the console if any.'''
 	def handle_URI(self, request, file_path):
 		if len(file_path) == 1:
 			return "index.html"
@@ -236,32 +245,35 @@ class server:
 
 		return file_path[1:]
 
+		'''Funtion to handle conditional GET request.'''
 	def conditional_req(self, request, conditional_get, new_connection, message):
 		conditional_get = time.strptime(conditional_get, "%a, %d %b %Y %H:%M:%S")
-		conditional_get = int(time.strftime("%s", conditional_get))
+		conditional_get = int(time.strftime("%s", conditional_get))#Get the time in seconds.
 		file_modified_date = os.path.getmtime(request['resource'])
-		file_modified_date = int(time.strftime("%s", time.localtime(file_modified_date)))
-		if "If-Modified-Since" in request.keys() and file_modified_date <= conditional_get:
-			print("No message")
-			response = message.handle_304(request)
-			response += "\r\n"
-			new_connection.send(response.encode())
-			return True
+		file_modified_date = int(time.strftime("%s", time.localtime(file_modified_date)))#modified in seconds.
+		if os.access(request['resource'], os.R_OK):
+			print("1234 5678")
+			if "If-Modified-Since" in request.keys() and file_modified_date <= conditional_get:
+				print("No message")
+				response = message.handle_304(request)
+				response += "\r\n"
+				new_connection.send(response.encode())
+				return True
 
-		elif "If-Unmodified-Since" in request.keys() and file_modified_date >= conditional_get:
-			print("modified")
-			response = message.handle_412(request)
-			response += "\r\n"
-			new_connection.send(response.encode())
-			return True
+			elif "If-Unmodified-Since" in request.keys() and file_modified_date >= conditional_get:
+				print("modified")
+				response = message.handle_412(request)
+				response += "\r\n"
+				new_connection.send(response.encode())
+				return True
 
 		return False
 
 
 	def get(self, request, new_connection):
-		message = Response(request)
+		message = Response(request)#Creates a response object
 		print("GET " + request['resource'] + " HTTP/1.1" )
-		if ("If-Modified-Since" or "If-Unmodified-Since")in request.keys():
+		if ("If-Modified-Since" or "If-Unmodified-Since") in request.keys(): #checks for conditional GET header
 			if "If-Modified-Since" in request.keys():
 				conditional_get = request['If-Modified-Since'][:-4]
 			elif "If-Unmodified-Since" in request.keys():
@@ -269,30 +281,34 @@ class server:
 			if self.conditional_req(request, conditional_get, new_connection, message):
 				return 
 
-		try:
-			file = open(request['resource'], "rb")
-		except:
+		#Checks file read permission
+		if not os.path.exists(request['resource']):
 			response = message.handle_4xx(request, 404)
 			new_connection.send(response.encode())
 			return 
+		if os.access(request['resource'], os.R_OK):
+			response = message.handle_200(request)
+			if not response:
+				response = message.handle_4xx(request, 415)
+				new_connection.send(response.encode())
+				return
 
-		response = message.handle_200(request)
-		if not response:
-			response = message.handle_4xx(request, 415)
+			file = open(request['resource'], "rb")
+			entity_len = os.stat(request['resource']).st_size	#gets length of file content
+			response += "Content-Length: " + str(entity_len) + "\r\n"
+			response += "\r\n"
+			response = response.encode()
+			new_connection.send(response)#first send the headers
+			l = new_connection.sendfile(file)# Used to send file 
+			file.close()
+			return
+		else:#if permission is denied to the file
+			response = message.handle_4xx(request, 403)
 			new_connection.send(response.encode())
 			return
 
-		entity_len = os.stat(request['resource']).st_size	
-		response += "Content-Length: " + str(entity_len) + "\r\n"
-		response += "\r\n"
-		response = response.encode()
-		new_connection.send(response)
-		l = new_connection.sendfile(file)
-		file.close()
-		return
-
 	def head(self, request, new_connection):
-		message = Response(request)
+		message = Response(request)#Creates a response object
 		print("HEAD " + request['resource'] + " HTTP/1.1" )
 		try:
 			file = open(request['resource'], "rb")
@@ -300,7 +316,7 @@ class server:
 			response = message.handle_4xx(request, 404)
 			new_connection.send(response.encode())
 			return 
-		entity_len = os.stat(request['resource']).st_size
+		entity_len = os.stat(request['resource']).st_size #gets length of file content
 		response = message.handle_200(request)
 		response += "Content-Length: " + str(entity_len) + "\r\n"
 		response += "\r\n"
@@ -310,16 +326,16 @@ class server:
 		return 
 
 	def post(self, request, body, new_connection):
-		message = Response(request)
+		message = Response(request) #Creates a response object
 		print("POST " + request['resource'] + " HTTP/1.1" )
 		if request['Content-Type'] == "application/x-www-form-urlencoded":
 			if "&" in body:
-				body = body.split("&")
-			if os.path.exists("wr.txt"):
+				body = body.split("&") #Split various key-value pairs
+			if os.access("wr.txt", os.W_OK): #Check write permission
 				try:
-					file = open("wr.txt", "a")
+					file = open("wr.txt", "a") #open file to write key value pairs
 					for i in body:
-						key, value = i.split("=")
+						key, value = i.split("=") 
 						file.write(value)
 						file.write(",")
 
@@ -327,23 +343,10 @@ class server:
 				except OSError:
 					print("Some error occurred during execution of POST")
 				file.close()
-			else:
-				try:
-					file = open("wr.txt", "w")
-					for i in body:
-						key, value = i.split("=")
-						file.write(key)
-						file.write(",")
-					file.write("\r\n")
-					for i in body:
-						key, value = i.split("=")
-						file.write(value)
-						file.write(",")
-					file.write("\r\n")
-
-				except OSError:
-					print("Some error occurred during execution of POST")
-				file.close()
+			else: #Write permission denied
+				response = message.handle_4xx(request, 403)
+				new_connection.send(response.encode())
+				return
 
 		response = message.handle_201(request)
 		new_connection.send(response.encode())
@@ -351,39 +354,70 @@ class server:
 	def put(self, request, recv_message, new_connection):
 		message = Response(request)
 		print("PUT " + request['resource'] + " HTTP/1.1" )
-		if os.path.exists(request['resource']):
+		user_pass = getpass.getpass(prompt = "Enter password: ")
+		if user_pass != self.authorization_pass:
+			print("Authentication failed")
+			response = message.handle_4xx(request, 401)
+			new_connection.send(response.encode())
+			return 
+
+		status = 201
+		if os.path.exists(request['resource']): # if already exists status is 204 otherwise, 201
 			status = 204
-		else:
-			status = 201
 		try:
 			file = open(request['resource'], "wb")
-		except:
+			file.write(recv_message)
+			print("written")
+			file.close()
+			if status == 201:
+				response = message.handle_201(request)
+			elif status == 204:
+				response = message.handle_204(request)
+				
+			new_connection.send(response.encode())
+		except: #permission denied
 			print("The file could not be created/updated")
-		file.write(recv_message)
-		file.close()
-		if status == 201:
-			response = message.handle_201(request)
-		elif status == 204:
-			response = message.handle_204(request)
-		new_connection.send(response.encode())
+			response = message.handle_4xx(request, 403)
+			new_connection.send(response.encode())
+			return
+		
 		return
 
 	def delete(self, request, new_connection):
 		message = Response(request)
 		print("DELETE " + request['resource'] + " HTTP/1.1" )
 		if os.path.exists(request['resource']):
-			try:
-				os.remove(request['resource'])
-			except IsADirectoryError:
-				shutil.rmtree(request['resource'])
-			response = message.handle_204(request)
-			new_connection.send(response.encode())
+			if os.access(request['resource'], os.W_OK) and os.access(request['resource'], os.R_OK):#check read-write permissions
+				''' Enter the password on the console to authenticate'''
+				user_pass = getpass.getpass(prompt = "Enter password: ")
+				if user_pass != self.authorization_pass:
+					#If wrong password => Unauthorized error
+					response = message.handle_4xx(request, 401)
+					new_connection.send(response.encode())
+					return 
+				try:
+					os.remove(request['resource'])#delete the file
+				except IsADirectoryError: # If its not he file but a directory : delete the whole directory.
+					shutil.rmtree(request['resource'])
+				response = message.handle_204(request)
+				new_connection.send(response.encode())
+			else:
+				response = message.handle_4xx(request, 403)
+				new_connection.send(response.encode())
+				return
 		else:
 			response = message.handle_4xx(request, 404)
 			new_connection.send(response.encode())
 
+	def trace(self, request, content, new_connection):
+		''' Sends the request packet as it is in the body '''
+		response = "HTTP/1.1 200 Ok\r\n"
+		response += "Content-Type: message/http\r\n"
+		response += "Date: " + curr_time() + "\r\n"
+		response += "Connection: " + request['connection'] + "\r\n\r\n"
+		new_connection.send(response.encode())
+		new_connection.send(content)
+
 if __name__ == '__main__':
-	host = "127.0.0.1"
-	port = 5677
-	address = (host, port)
-	new_server = server(address)
+	address = config.address#obtain address from the configuration file
+	new_server = server(address) # Create server object
